Updating objectTracker and COM error reader to use integers

+++ openCV/ESP32_Scripts/readCameraError/readCameraError.ino
@@ -7,8 +7,8 @@ void loop() {
   if (Serial.available()) {
     String msg = Serial.readStringUntil('\n');
     
-    float ex, ey;
-    sscanf(msg.c_str(), "%f,%f", &ex, &ey);
-    Serial.printf("Error x-axis: %.2f, Error y-axis: %.2f\n", ex, ey); // temp
+    int ex, ey;
+    sscanf(msg.c_str(), "%d,%d", &ex, &ey);
+    Serial.printf("Error x-axis: %d, Error y-axis: %d\n", ex, ey); // temp
   }
 }
